+++ patches/Accessibility/strings_de.cpp
@@ -453,6 +453,8 @@ const char* Get(Id id) {
         case Id::ModSettingStateOn:           return "an";
         case Id::ModSettingStateOff:          return "aus";
         case Id::FmtModSettingOption:         return "%s: %s";
+        case Id::ModSettingCueVolume:         return "Lautst\xe4rke der Hinweist\xf6ne";
+        case Id::FmtModSettingSlider:         return "%s: %d Prozent";
 
         case Id::ModSettingAudioGlossary:           return "Audio-Glossar";
         case Id::ModSettingsAudioGlossaryOpened:    return "Audio-Glossar ge\xf6""ffnet";